Fix analog pin resolution

+++ controllers/generic_controller/lib/AnalogPin/AnalogPin.cpp
@@ -4,7 +4,7 @@ bool AnalogPin::readable() {
   return _enabled;
 };
 
-uint8_t AnalogPin::readIfAllowed() {
+uint16_t AnalogPin::readIfAllowed() {
   if (readable()) {
     return analogRead(_physicalPin);
   } else {
