+++ controllers/generic_controller/lib/AnalogPin/AnalogPin.h
@@ -8,7 +8,7 @@ class AnalogPin: public OtherPin {
     AnalogPin() {};
     AnalogPin(bool enabled, uint8_t physicalPin) : OtherPin(enabled, physicalPin) {};
     bool readable();
-    uint8_t readIfAllowed();
+    uint16_t readIfAllowed();
 };
 
 #endif
